\name{RbioRXN-package}
\alias{RbioRXN-package}
\alias{RbioRXN}
\docType{package}
\title{
RbioRXN - Download/process Rhea, MetaCyc, KEGG and UniPathway biochemical reaction
}
\description{
RbioRXN is an R package that facilitates gaining and processing biochemical reaction such as Rhea, MetaCyc, KEGG and UniPathway. The package provides the functions to download and parse raw data, instantiate generic reaction, and check mass balance. The package aims to construct an integrated metabolic network and genome-scale metabolic model.
}
\details{
\tabular{ll}{
Package: \tab RbioRXN\cr
Type: \tab Package\cr
Version: \tab 1.5\cr
Date: \tab 2015-01-14\cr
License: \tab GPL (>=2)\cr
}
}
\author{
Byoungnam Min, Byeonghyuk Park, Kyoung Heon Kim and In-Geol Choi \cr
Maintainer: Byoungnam Min <mbnmbn00@gmail.com>
}
\references{
http://compbio.korea.ac.kr
}
\keyword{ KEGG }
\keyword{ Biochemical reaction database }
\keyword{ MetaCyc }
\keyword{ Rhea }
\keyword{ UniPathway }
\keyword{ Generic reaction }
\keyword{ Instantiation }
